DiscordCurrencyBot
C+urrently being updated but stores money in a json file and has an inventory system that is currently not in use.

Current commands are:
.ban @person#0001 [reason]
.kick @person#0001 [reason]
note that the user has to be pinged. and that there are allot of bugs that can crash the bot with this.

|| CURRENCY ||

.start - creates a json profile for the person so you can then store info about them.
.give @player#0001 Money Amount. 
(Floats can be stored too. I think)

That's it for now. Although this will be definitely updated.